requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
blinker==1.6.3
Brotli==1.0.9
capstone==5.0.1
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colored-traceback==0.3.0
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
discord==2.3.2
discord.py==2.3.2
distro==1.7.0
distro-info==1.1+ubuntu0.1
dotenv-cli==2.2.0
exceptiongroup==1.2.1
Flask==3.0.0
fonttools==4.29.1
frozenlist==1.4.0
fs==2.4.12
gast==0.5.2
git-filter-repo==2.38.0
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.4
importlib-metadata==4.6.4
intervaltree==3.1.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.3
lz4==3.1.3+dfsg
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.5.1
mdurl==0.1.2
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
netifaces==0.11.0
noise==1.2.2
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
orjson==3.9.7
outcome==1.3.0.post0
packaging==23.2
paramiko==3.3.1
Pillow==10.1.0
plumbum==1.8.2
ply==3.11
psutil==5.9.5
pwntools==4.11.0
pycparser==2.21
pyelftools==0.30
pyglet==1.5.27
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyradios==2.1.0
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-vlc==3.0.20123
pythran==0.10.0
pytz==2022.1
PyVirtualDisplay==3.0
PyYAML==5.4.1
requests==2.31.0
rich==13.7.1
ROPGadget==7.4
rpyc==5.3.1
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.1
sympy==1.9
systemd-python==234
tomli==2.0.1
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unicorn==2.0.1.post1
urllib3==2.0.4
wadllib==1.3.6
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.0
wsproto==1.2.0
yarl==1.9.2
zipp==1.0.0
